+++ plugins/harlequin_sqlite/__init__.py
@@ -1,0 +1,2 @@
+from harlequin_sqlite.adapter import HarlequinSqliteAdapter as HarlequinSqliteAdapter
+from harlequin_sqlite.cli_options import SQLITE_OPTIONS as SQLITE_OPTIONS